MAJOR: runzeus.sh: try to re-download the license if provided over HTTP and start up is in error
For *some* reasons, when I use docker-compose to start a vtm AND I
provide the license over an HTTP URL using a real hostname (resolved
using DNS), then it may happens that the hostname can't be resolved at
first try (I know, this is really weird....).
In order to workaround this issue, I simply retry to download the
license in the loop handing the initial configuration error.